Make separate class for session pool
